Upgrade dependencies
fix: Refactor Navbar routing for React Router v7
