+++ website/jest.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  setupFiles: ['<rootDir>/src/setupFiles.js'],
+};
