+++ website/package.json
@@ -3,15 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "apexcharts": "^3.25.0",
-    "react": "^16.13.1",
-    "react-apexcharts": "^1.3.7",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "apexcharts": "^4.5.0",
+    "react": "^19.1.0",
+    "react-apexcharts": "^1.7.0",
+    "react-dom": "^19.1.0",
+    "react-router": "^7.5.1",
+    "react-scripts": "^5.0.1",
     "react-scroll": "^1.8.1",
     "recharts": "^2.0.8"
   },
@@ -41,10 +41,10 @@
     ]
   },
   "devDependencies": {
-    "lint-staged": "^11.1.1",
-    "prettier": "^2.3.1",
-    "stylelint": "^13.11.0",
-    "stylelint-config-standard": "^20.0.0"
+    "lint-staged": "^15.5.1",
+    "prettier": "^3.5.3",
+    "stylelint": "^16.18.0",
+    "stylelint-config-standard": "^38.0.0"
   },
   "lint-staged": {
     "*.(js|html)": "prettier --write",
